refactor: split server to a separated file
